+++ src/calculator/Operations.java
@@ -1,24 +1,25 @@
 package calculator;
 
+import java.util.LinkedList;
 import java.util.List;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.stream.Collectors;
 
 public class Operations {
-    public void suma(List<Integer> listOne, List<Integer> listTwo){
+    public void suma(List<Double> listOne, List<Double> listTwo){
         AtomicInteger indice = new AtomicInteger();
 
         System.out.println("SUMA: " + listOne.stream().map(element->element + listTwo.get(indice.getAndIncrement()))
                 .collect(Collectors.toList()));
     }
 
-    public void resta(List<Integer> listOne, List<Integer> listTwo){
+    public void resta(List<Double> listOne, List<Double> listTwo){
         AtomicInteger indice = new AtomicInteger();
         System.out.println("RESTA: "+listOne.stream().map(element->element - listTwo.get(indice.getAndIncrement()))
                 .collect(Collectors.toList()));
     }
 
-    public void multiplication(List<Integer> listOne, List<Integer> listTwo){
+    public void multiplication(List<Double> listOne, List<Double> listTwo){
         AtomicInteger indice = new AtomicInteger();
         System.out.println("MULTIPLICACIÓN: "+listOne.stream().map(element->element * listTwo.get(indice.getAndIncrement()))
                 .collect(Collectors.toList()));
@@ -26,7 +27,7 @@ public class Operations {
     }
 
 
-    public void division(List<Integer> listOne, List<Integer> listTwo){
+    public void division(List<Double> listOne, List<Double> listTwo){
 
         AtomicInteger indice = new AtomicInteger();
         System.out.println("DIVISIÓN: "+listTwo.stream()
@@ -40,7 +41,7 @@ public class Operations {
                 .collect(Collectors.toList()));
     }
 
-    public void modulo(List<Integer> listOne, List<Integer> listTwo){
+    public void modulo(List<Double> listOne, List<Double> listTwo){
         AtomicInteger indice = new AtomicInteger();
         System.out.println("MODULO: "+listTwo.stream()
                 .map(element-> {
